+++ CHANGELOG.md
@@ -52,6 +52,7 @@
 - Fixed validation of options of resource operations in `pcs resource create`
   ([rhbz#1390071])
 - Fixed silent omission of duplicate options ([rhbz#1390066])
+- Added more validation for resource agent names ([rhbz#1387670])
 
 [ghissue#124]: https://github.com/ClusterLabs/pcs/issues/124
 [ghissue#125]: https://github.com/ClusterLabs/pcs/issues/125
@@ -65,6 +66,7 @@
 [rhbz#1378107]: https://bugzilla.redhat.com/show_bug.cgi?id=1378107
 [rhbz#1382004]: https://bugzilla.redhat.com/show_bug.cgi?id=1382004
 [rhbz#1382597]: https://bugzilla.redhat.com/show_bug.cgi?id=1382597
+[rhbz#1387670]: https://bugzilla.redhat.com/show_bug.cgi?id=1387670
 [rhbz#1389443]: https://bugzilla.redhat.com/show_bug.cgi?id=1389443
 [rhbz#1389501]: https://bugzilla.redhat.com/show_bug.cgi?id=1389501
 [rhbz#1389941]: https://bugzilla.redhat.com/show_bug.cgi?id=1389941